Working on the FK with tables problem.
This approach uses the nested tables approach recommended in
stackoverflow. The Play framework docs are excreble here, and provide
only illusory guidance. It is probable that the simple way to do it is
the old fashioned php'y way of "read data, stuff it here", as opposed to
fine, fine ORM queries. This is trial 2 in the lists; on trial 3 the ORM
shall be thrown out unceremoniously and replaced with a SQL query and
some hand-jiggery to build the forms.

Current Mood: No Abstractions Without Documentation!

... it's midnight:45